requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.8.30
click==8.1.7
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.0.1
rcs==1.0.18
sniffio==1.3.1
typing_extensions==4.12.2
